Fix build for netcoreapp3.0, which doesn't have SerializeToStreamAsync with cancellation.
